+++ driven/storage/adapter_events.go
@@ -4,6 +4,7 @@ import (
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/bson/primitive"
 	"groups/core/model"
+	"time"
 )
 
 // FindAdminGroupsForEvent Finds all groups for an event where the user is admin
@@ -78,7 +79,7 @@ func (sa *Adapter) FindAdminGroupsIDs(context TransactionContext, clientID strin
 	}
 	var result []aggregator
 
-	err := sa.db.events.AggregateWithContext(context, pipeline, &result, nil)
+	err := sa.db.groupMemberships.AggregateWithContext(context, pipeline, &result, nil)
 	if err != nil {
 		return nil, err
 	}
@@ -90,11 +91,12 @@ func (sa *Adapter) FindAdminGroupsIDs(context TransactionContext, clientID strin
 	return nil, nil
 }
 
-// UpdateGroupsForEvent Updates group mappings for an event
-func (sa *Adapter) UpdateGroupsForEvent(context TransactionContext, clientID string, current *model.User, eventID string, groupIDs []string) ([]string, error) {
-	var reult []string
+// UpdateGroupMappingsForEvent Updates group mappings for an event
+func (sa *Adapter) UpdateGroupMappingsForEvent(context TransactionContext, clientID string, current *model.User, eventID string, groupIDs []string) ([]string, error) {
+	var result []string
 
 	wrapper := func(context TransactionContext) error {
+		// 1. Construct mappings for lookups
 		adminIDMappings := map[string]bool{}
 		adminGroupIDs, err := sa.FindAdminGroupsIDs(context, clientID, current)
 		if err != nil {
@@ -102,6 +104,72 @@ func (sa *Adapter) UpdateGroupsForEvent(context TransactionContext, clientID str
 		}
 		for _, groupID := range adminGroupIDs {
 			adminIDMappings[groupID] = true
+		}
+
+		currentAdminIDMappings := map[string]bool{}
+		currentAdminGroupIDs, err := sa.FindAdminGroupsForEvent(context, clientID, current, eventID)
+		if err != nil {
+			return err
+		}
+		for _, groupID := range currentAdminGroupIDs {
+			currentAdminIDMappings[groupID] = true
+		}
+
+		newGroupIDsMapping := map[string]bool{}
+		for _, groupID := range groupIDs {
+			newGroupIDsMapping[groupID] = true
+		}
+
+		// 2. Construct mappings for remove
+		var groupIDsForRemove []string
+		for _, groupID := range groupIDs {
+			if !currentAdminIDMappings[groupID] && adminIDMappings[groupID] {
+				groupIDsForRemove = append(groupIDsForRemove, groupID)
+			}
+		}
+		if len(groupIDsForRemove) > 0 {
+			_, err := sa.db.events.DeleteManyWithContext(context, bson.D{
+				{Key: "event_id", Value: eventID},
+				{Key: "group_id", Value: bson.M{"$in": groupIDsForRemove}},
+				{Key: "client_id", Value: clientID},
+			}, nil)
+			if err != nil {
+				return err
+			}
+		}
+
+		for _, groupID := range groupIDs {
+			if adminIDMappings[groupID] {
+				result = append(result, groupID)
+			}
+		}
+
+		var eventsForAdd []interface{}
+		for _, groupID := range groupIDs {
+			if _, ok := currentAdminIDMappings[groupID]; !ok {
+				if _, innerOK := adminIDMappings[groupID]; innerOK {
+					eventsForAdd = append(eventsForAdd, model.Event{
+						ClientID: clientID,
+						GroupID:  groupID,
+						EventID:  eventID,
+						Creator: &model.Creator{
+							UserID: current.ID,
+							Name:   current.Name,
+							Email:  current.Email,
+						},
+						DateCreated: time.Now(),
+					})
+				}
+			}
+			if !currentAdminIDMappings[groupID] && adminIDMappings[groupID] {
+
+			}
+		}
+		if len(eventsForAdd) > 0 {
+			_, err := sa.db.events.InsertManyWithContext(context, eventsForAdd, nil)
+			if err != nil {
+				return err
+			}
 		}
 
 		return nil
@@ -117,5 +185,5 @@ func (sa *Adapter) UpdateGroupsForEvent(context TransactionContext, clientID str
 		return nil, err
 	}
 
-	return reult, nil
+	return result, nil
 }