+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Retrieve group ids by event Id [#416](https://github.com/rokwire/groups-building-block/issues/416)
+- Update the set of groups that event is published to [#417](https://github.com/rokwire/groups-building-block/issues/417)
 
 ## [1.29.0] - 2024-02-06
 ### Changed